requirements.txt
# Institutional-Grade Signal Scoring Bot Dependencies

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
pytz>=2023.3

# API Clients
ccxt>=4.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Configuration & Environment
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Financial Analysis
TA-Lib>=0.4.25
scipy>=1.10.0
scikit-learn>=1.3.0

# Async & Performance
asyncio-mqtt>=0.13.0
uvloop>=0.17.0

# Logging & Monitoring
loguru>=0.7.0
rich>=13.0.0

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Web Framework (for future API endpoints)
fastapi>=0.100.0
uvicorn>=0.23.0

# Database (for storing historical data)
sqlalchemy>=2.0.0
alembic>=1.11.0

# Utilities
python-dateutil>=2.8.0
websocket-client>=1.6.0